WIP: Refactor hav helper into a decorator
