+++ app/controllers/concerns/application_controller/request_preprocessing.rb
@@ -72,8 +72,10 @@ module Concerns::ApplicationController::RequestPreprocessing
     Rails.logger.info("Request URL: #{request.url}")
   end
 
+  # deletable
   def set_default_nav_context
     @context = {}
+    nav_link_builder.context = {}
   end
 
   # Checks that the subdomain's community exists and sets current_community.
