+++ app/views/layouts/application.html.erb
@@ -33,7 +33,7 @@
           <td class="main-nav">
             <% if user_signed_in? %>
               <ul class="nav">
-                <% nav_items.each do |item| %>
+                <% nav_link_builder.nav_items.each do |item| %>
                   <li class="<%= item[:active] ? 'active' : '' %>">
                     <%= link_to(t("nav_links.main.#{item[:name]}"),
                       item[:path], class: "icon-bar") %>
@@ -45,12 +45,12 @@
         </tr>
       </table>
     </nav>
-    <% if @context[:subsection] && (items = subnav_items).size > 1 %>
+    <% if @context[:subsection] && (items = nav_link_builder.subnav_items).size > 1 %>
       <nav class="subnav hidden-xs">
         <ul class="nav nav-tabs" role="tablist">
           <% items.each do |item| %>
             <li role="presentation" class="<%= item[:active] ? 'active' : '' %>">
-              <%= nav_link(item, tab: true) %>
+              <%= nav_link_builder.nav_link(item, tab: true) %>
             </li>
           <% end %>
         </ul>
@@ -74,8 +74,8 @@
         <% if user_signed_in? %>
           <div class="collapse navbar-collapse" id="mobile-nav">
             <ul class="nav navbar-nav mobile-nav">
-              <% nav_items.each do |item| %>
-                <% if (sub_items = subnav_items(item[:name])).size > 1 %>
+              <% nav_link_builder.nav_items.each do |item| %>
+                <% if (sub_items = nav_link_builder.subnav_items(item[:name])).size > 1 %>
                   <li class="dropdown">
                     <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button"
                       aria-haspopup="true" aria-expanded="false">
@@ -85,12 +85,12 @@
                     </a>
                     <ul class="dropdown-menu">
                       <% sub_items.each do |sub_item| %>
-                        <li><%= nav_link(sub_item) %></li>
+                        <li><%= nav_link_builder.nav_link(sub_item) %></li>
                       <% end %>
                     </ul>
                   </li>
                 <% else %>
-                  <li><%= nav_link(item) %></li>
+                  <li><%= nav_link_builder.nav_link(item) %></li>
                 <% end %>
               <% end %>
               <%= render("layouts/personal_dropdown") %>
